+++ crates/orchestrator/src/tests/common/mod.rs
@@ -1,66 +1,23 @@
 pub mod constants;
 
 use std::collections::HashMap;
-use std::sync::Arc;
 
 use ::uuid::Uuid;
 use aws_config::Region;
-use constants::*;
-use da_client_interface::MockDaClient;
 use mongodb::Client;
-use prover_client_interface::MockProverClient;
 use rstest::*;
 use serde::Deserialize;
-use settlement_client_interface::MockSettlementClient;
-use starknet::providers::jsonrpc::HttpTransport;
-use starknet::providers::JsonRpcClient;
-use url::Url;
 
-use crate::config::Config;
 use crate::data_storage::aws_s3::config::{AWSS3ConfigType, S3LocalStackConfig};
 use crate::data_storage::aws_s3::AWSS3;
-use crate::data_storage::{DataStorage, DataStorageConfig, MockDataStorage};
+use crate::data_storage::{DataStorage, DataStorageConfig};
 use crate::database::mongodb::config::MongoDbConfig;
 use crate::database::mongodb::MongoDb;
-use crate::database::{DatabaseConfig, MockDatabase};
+use crate::database::DatabaseConfig;
 use crate::jobs::types::JobStatus::Created;
 use crate::jobs::types::JobType::DataSubmission;
 use crate::jobs::types::{ExternalId, JobItem};
 use crate::queue::job_queue::{JOB_PROCESSING_QUEUE, JOB_VERIFICATION_QUEUE};
-use crate::queue::MockQueueProvider;
-
-pub async fn init_config(
-    rpc_url: Option<String>,
-    database: Option<MockDatabase>,
-    queue: Option<MockQueueProvider>,
-    da_client: Option<MockDaClient>,
-    prover_client: Option<MockProverClient>,
-    settlement_client: Option<MockSettlementClient>,
-    storage_client: Option<MockDataStorage>,
-) -> Config {
-    let _ = tracing_subscriber::fmt().with_max_level(tracing::Level::INFO).with_target(false).try_init();
-
-    let rpc_url = rpc_url.unwrap_or(MADARA_RPC_URL.to_string());
-    let database = database.unwrap_or_default();
-    let queue = queue.unwrap_or_default();
-    let da_client = da_client.unwrap_or_default();
-    let prover_client = prover_client.unwrap_or_default();
-    let settlement_client = settlement_client.unwrap_or_default();
-    let storage_client = storage_client.unwrap_or_default();
-
-    // init starknet client
-    let provider = JsonRpcClient::new(HttpTransport::new(Url::parse(rpc_url.as_str()).expect("Failed to parse URL")));
-
-    Config::new(
-        Arc::new(provider),
-        Box::new(da_client),
-        Box::new(prover_client),
-        Box::new(settlement_client),
-        Box::new(database),
-        Box::new(queue),
-        Box::new(storage_client),
-    )
-}
 
 #[fixture]
 pub fn default_job_item() -> JobItem {